requirements.txt
telethon>=1.30.0

asgiref>=3.7.0
Pillow>=10.0.0 

whitenoise

# requirements.txt
# Все зависимости проекта SalesAI

# ============================================
# DJANGO
# ============================================
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.11.2
django-allauth>=0.65.0
# ============================================
# БАЗА ДАННЫХ
# ============================================
psycopg2-binary==2.9.9
pgvector==0.2.5

# ============================================
# ПЕРЕМЕННЫЕ ОКРУЖЕНИЯ
# ============================================
python-dotenv==1.0.0

# ============================================
# REST API
# ============================================
djangorestframework==3.14.0

# ============================================
# CELERY (асинхронные задачи)
# ============================================
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-redis==5.4.0

# ============================================
# OPENAI И RAG
# ============================================
openai==1.12.0
pypdf==3.17.4
python-docx==1.1.0

# ============================================
# РАБОТА С ФАЙЛАМИ
# ============================================
Pillow==10.2.0
python-magic==0.4.27

# ============================================
# HTTP КЛИЕНТЫ
# ============================================
requests==2.31.0

# ============================================
# БОТЫ (опционально)
# ============================================
python-telegram-bot==20.7
twilio==8.11.1

# ============================================
# УТИЛИТЫ
# ============================================
python-dateutil==2.8.2
pytz==2024.1

# ============================================
# WSGI СЕРВЕР ДЛЯ PRODUCTION
# ============================================
gunicorn==21.2.0

# ============================================
# ТЕСТИРОВАНИЕ (опционально)
# ============================================
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0